REF: collect DataFrame.drop, Series.drop tests (#32994)
